requirements.txt
Flask==3.1.2
Werkzeug==3.1.4
Jinja2==3.1.6
click==8.1.8
itsdangerous==2.2.0
python-dotenv==1.0.1
mysql-connector-python==9.1.0